--- database/src/main/java/by/borzdykooa/entity/OrderingMedicine.java
@@ -1,35 +1,0 @@
-package by.borzdykooa.entity;
-
-import by.borzdykooa.entity.helpers.IdEntity;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
-@AllArgsConstructor
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@Entity
-@Table(name = "ordering_medicine", schema = "online_pharmacy")
-public class OrderingMedicine extends IdEntity {
-
-    @ManyToOne
-    @JoinColumn(name = "ordering_id", nullable = false)
-    private Ordering ordering;
-
-    @ManyToOne
-    @JoinColumn(name = "medicine_id", nullable = false)
-    private Medicine medicine;
-
-    @Column(name = "quantity", nullable = false)
-    private Long quantity;
-}
