--- database/src/main/java/by/borzdykooa/entity/enums/Status.java
@@ -1,16 +1,0 @@
-package by.borzdykooa.entity.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum Status {
-
-    DONE("выполнен"),
-    PROCESSED("обрабатывается");
-
-    private String description;
-
-    Status(String description) {
-        this.description = description;
-    }
-}
