--- database/src/main/java/by/borzdykooa/entity/helpers/IdEntity.java
@@ -1,21 +1,0 @@
-package by.borzdykooa.entity.helpers;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-
-@Getter
-@Setter
-@ToString
-@MappedSuperclass
-public class IdEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-}
