+++ database/src/main/java/com/borzdykooa/entity/OrderingMedicine.java
@@ -1,0 +1,35 @@
+package com.borzdykooa.entity;
+
+import com.borzdykooa.entity.helpers.IdEntity;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+@AllArgsConstructor
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@Entity
+@Table(name = "ordering_medicine", schema = "online_pharmacy")
+public class OrderingMedicine extends IdEntity <Long>{
+
+    @ManyToOne
+    @JoinColumn(name = "ordering_id", nullable = false)
+    private Ordering ordering;
+
+    @ManyToOne
+    @JoinColumn(name = "medicine_id", nullable = false)
+    private Medicine medicine;
+
+    @Column(name = "quantity", nullable = false)
+    private Long quantity;
+}
