+++ database/src/main/java/com/borzdykooa/entity/enums/UserRole.java
@@ -1,0 +1,15 @@
+package com.borzdykooa.entity.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum  UserRole {
+    CLIENT ("клиент"),
+    ADMIN ("администратор");
+
+    private String description;
+
+    UserRole(String description) {
+        this.description = description;
+    }
+}
