+++ database/src/main/java/com/borzdykooa/entity/helpers/FullName.java
@@ -1,0 +1,26 @@
+package com.borzdykooa.entity.helpers;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@Embeddable
+public class FullName {
+
+    @Column(name = "last_name", nullable = false)
+    private String lastName;
+
+    @Column(name = "first_name", nullable = false)
+    private String firstName;
+
+    @Column(name = "patronymic", nullable = false)
+    private String patronymic;
+}
