+++ database/src/main/java/com/borzdykooa/entity/helpers/IdEntity.java
@@ -1,0 +1,22 @@
+package com.borzdykooa.entity.helpers;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+import java.io.Serializable;
+
+@Getter
+@Setter
+@ToString
+@MappedSuperclass
+public class IdEntity <PK extends Serializable> {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private PK id;
+}
