+++ database/src/main/java/com/borzdykooa/manager/SessionFactoryManager.java
@@ -1,0 +1,16 @@
+package com.borzdykooa.manager;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import org.hibernate.SessionFactory;
+import org.hibernate.cfg.Configuration;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class SessionFactoryManager {
+
+    private static final SessionFactory SESSION_FACTORY = new Configuration().configure().buildSessionFactory();
+
+    public static SessionFactory getSessionFactory() {
+        return SESSION_FACTORY;
+    }
+}
