+++ database/src/test/resources/hibernate.cfg.xml
@@ -16,14 +16,39 @@
         <!-- DB schema will be updated if needed -->
         <property name="hbm2ddl.auto">create-drop</property>
 
-        <mapping class="by.borzdykooa.entity.User"/>
-        <mapping class="by.borzdykooa.entity.Admin"/>
-        <mapping class="by.borzdykooa.entity.Client"/>
-        <mapping class="by.borzdykooa.entity.Prescription"/>
-        <mapping class="by.borzdykooa.entity.Review"/>
-        <mapping class="by.borzdykooa.entity.Medicine"/>
-        <mapping class="by.borzdykooa.entity.Ordering"/>
-        <mapping class="by.borzdykooa.entity.OrderingMedicine"/>
-        <mapping class="by.borzdykooa.entity.PharmacyGroup"/>
+        <mapping class="com.borzdykooa.entity.User"/>
+        <mapping class="com.borzdykooa.entity.Admin"/>
+        <mapping class="com.borzdykooa.entity.Client"/>
+        <mapping class="com.borzdykooa.entity.Prescription"/>
+        <mapping class="com.borzdykooa.entity.Review"/>
+        <mapping class="com.borzdykooa.entity.Medicine"/>
+        <mapping class="com.borzdykooa.entity.Ordering"/>
+        <mapping class="com.borzdykooa.entity.OrderingMedicine"/>
+        <mapping class="com.borzdykooa.entity.PharmacyGroup"/>
+        <mapping class="com.borzdykooa.entity.SaleInfo"/>
     </session-factory>
 </hibernate-configuration>
+
+
+<!--<property name="connection.url">jdbc:postgresql://localhost:5432/online_pharmacy</property>-->
+<!--<property name="connection.driver_class">org.postgresql.Driver</property>-->
+<!--<property name="connection.username">postgres</property>-->
+<!--<property name="connection.password">pass</property>-->
+<!--<property name="dialect">org.hibernate.dialect.PostgreSQL95Dialect</property>-->
+<!--<property name="show_sql">true</property>-->
+<!--<property name="format_sql">true</property>-->
+
+        <!--&lt;!&ndash; DB schema will be updated if needed &ndash;&gt;-->
+        <!--&lt;!&ndash; <property name="hbm2ddl.auto">update</property> &ndash;&gt;-->
+<!--<mapping class="com.borzdykooa.entity.User"/>-->
+<!--<mapping class="com.borzdykooa.entity.Admin"/>-->
+<!--<mapping class="com.borzdykooa.entity.Client"/>-->
+<!--<mapping class="com.borzdykooa.entity.Prescription"/>-->
+<!--<mapping class="com.borzdykooa.entity.Review"/>-->
+<!--<mapping class="com.borzdykooa.entity.Medicine"/>-->
+<!--<mapping class="com.borzdykooa.entity.Ordering"/>-->
+<!--<mapping class="com.borzdykooa.entity.OrderingMedicine"/>-->
+<!--<mapping class="com.borzdykooa.entity.PharmacyGroup"/>-->
+<!--<mapping class="com.borzdykooa.entity.SaleInfo"/>-->
+        <!--</session-factory>-->
+        <!--</hibernate-configuration>-->
