--- service/src/main/java/by/borzdykooa/service/MedicineService.java
@@ -1,20 +1,0 @@
-package by.borzdykooa.service;
-
-import by.borzdykooa.dao.MedicineDao;
-import by.borzdykooa.entity.Medicine;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class MedicineService {
-
-    private static final MedicineService INSTANCE = new MedicineService();
-
-    public Medicine getMedicineById() {
-        return MedicineDao.getInstance().findMedicineById();
-    }
-
-    public static MedicineService getInstance() {
-        return INSTANCE;
-    }
-}
